ct: move snapshot and version into level zero stm
