+++ src/v/cloud_topics/level_zero/stm/ctp_stm_state.cc
@@ -10,7 +10,6 @@
 
 #include "cloud_topics/level_zero/stm/ctp_stm_state.h"
 
-#include "cloud_topics/dl_snapshot.h"
 #include "model/fundamental.h"
 
 #include <algorithm>
