+++ src/v/cloud_topics/level_zero/stm/dl_version.h
